fixed auth
