+++ src/store/constants.js
@@ -8,9 +8,4 @@ export const FIREBASE_CONFIG = {
 };
 
 export const SIGNUP_SUCCESS = "SIGNUP_SUCCESS";
-export const SIGNUP_ERROR = "SIGNUP_ERROR";
 export const SIGNIN_SUCCESS = "SIGNIN_SUCCESS";
-export const SIGNIN_ERROR = "SIGNIN_ERROR";
-export const CLEAR_SIGNUP_ERROR = "CLEAR_SIGNUP_ERROR";
-export const CLEAR_SIGNUP_MESSAGE = "CLEAR_SIGNUP_MESSAGE";
-export const CLEAR_SIGNIN_ERROR = "CLEAR_SIGNIN_ERROR";
