+++ src/view/Auth/SignUp/SignUp.jsx
@@ -1,0 +1,206 @@
+import React from "react";
+import firebase from "firebase";
+import { useDispatch } from "react-redux";
+import { Link } from "react-router-dom";
+import { signUpSuccess } from "./../../../store/actions/auth.action";
+import {
+  VALIDATORS_PASSWORD,
+  VALIDATORS_EMAIL,
+  isValidateField,
+} from "./../services/validate.service";
+import "./../Auth.scss";
+
+const SignUp = () => {
+  const [validateError, setValidateError] = React.useState({
+    errorEmail: [],
+    errorPassword: [],
+    errorRegister: [],
+  });
+  const [inputs, setInputs] = React.useState({
+    email: "",
+    password: "",
+    confirmPassword: "",
+  });
+
+  const dispatch = useDispatch();
+
+  const signUp = (credentials) => {
+    const { email, password, confirmPassword } = credentials;
+    const db = firebase.firestore().collection("users");
+    firebase
+      .auth()
+      .createUserWithEmailAndPassword(email, password)
+      .then((res) => {
+        db.doc(res.user.uid).set({
+          email,
+          password,
+        });
+        dispatch(
+          signUpSuccess({
+            isAuth: true,
+            userId: res.user.uid,
+            email,
+            password,
+            confirmPassword,
+          })
+        );
+      })
+      .catch((error) => {
+        setValidateError({
+          errorEmail: [],
+          errorPassword: [],
+          errorRegister: [error.message],
+        });
+      });
+  };
+
+  const hasInputsValue = () => {
+    const { email, password, confirmPassword } = inputs;
+    return email === "" || password === "" || confirmPassword === "";
+  };
+
+  React.useEffect(() => {
+    return () => {
+      setValidateError({
+        errorEmail: [],
+        errorPassword: [],
+        errorRegister: [],
+      });
+    };
+  }, []);
+
+  const handleChange = (e) => {
+    setInputs({ ...inputs, [e.target.name]: e.target.value });
+  };
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    const errorFieldPassword = isValidateField(VALIDATORS_PASSWORD, inputs);
+    const errorFieldEmail = isValidateField(VALIDATORS_EMAIL, inputs);
+    const isErrors = [...errorFieldEmail, ...errorFieldPassword].length;
+    if (isErrors) {
+      setValidateError({
+        errorEmail: [...errorFieldEmail],
+        errorPassword: [...errorFieldPassword],
+        errorRegister: [],
+      });
+    } else {
+      setValidateError({
+        errorEmail: [],
+        errorPassword: [],
+        errorRegister: [],
+      });
+      signUp(inputs);
+    }
+  };
+
+  const renderPasswordErrors = () =>
+    validateError.errorPassword.map((item) => (
+      <span key={item} className="error">
+        {item}
+      </span>
+    ));
+
+  const renderEmailErrors = () =>
+    validateError.errorEmail.map((item) => (
+      <span key={item} className="error">
+        {item}
+      </span>
+    ));
+
+  const renderRegisterErrors = () =>
+    validateError.errorRegister.map((item) => (
+      <span key={item} className="error">
+        {item}
+      </span>
+    ));
+
+  return (
+    <div className="wrapper">
+      <div className="wrapper-form">
+        <div className="title">
+          <h2 className="title-text">Register</h2>
+        </div>
+        <form className="sign-form" onSubmit={handleSubmit}>
+          <div className="sign-form_group">
+            <input
+              type="text"
+              name="email"
+              id="email"
+              className={`${
+                validateError.errorEmail.length && "sign-form_input_error"
+              } 
+                  sign-form_input
+              `}
+              onChange={handleChange}
+              required
+            />
+            <label className="sign-form_label" htmlFor="email">
+              Email
+            </label>
+            <div className={validateError.errorEmail.length && "info"}>
+              {renderEmailErrors()}
+            </div>
+          </div>
+
+          <div className="sign-form_group">
+            <input
+              type="password"
+              name="password"
+              id="password"
+              className={`${
+                validateError.errorPassword.length && "sign-form_input_error"
+              } 
+                  sign-form_input
+              `}
+              onChange={handleChange}
+              required
+            />
+            <label className="sign-form_label" htmlFor="password">
+              Password
+            </label>
+          </div>
+          <div className="sign-form_group">
+            <input
+              type="confirmPassword"
+              name="confirmPassword"
+              id="confirmPassword"
+              className={`${
+                validateError.errorPassword.length && "sign-form_input_error"
+              } 
+                  sign-form_input
+              `}
+              onChange={handleChange}
+              required
+            />
+            <label className="sign-form_label" htmlFor="confirmPassword">
+              Confirm Password
+            </label>
+            <div className={validateError.errorPassword.length && "info"}>
+              {renderPasswordErrors()}
+            </div>
+            <div className={validateError.errorRegister.length && "info"}>
+              {renderRegisterErrors()}
+            </div>
+          </div>
+          <div className="sign-form_group">
+            <button
+              className="sign-form_button"
+              onClick={handleSubmit}
+              disabled={hasInputsValue()}
+            >
+              Register
+            </button>
+          </div>
+          <div className="sign-form_group">
+            <Link className="sign-form_link" to="/signIn">
+              Already have an accaunt?
+            </Link>
+          </div>
+        </form>
+      </div>
+    </div>
+  );
+};
+
+export default SignUp;
